requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
soundcard>=0.4.2
numpy>=1.24.0
python-dotenv>=1.0.0
coloredlogs>=15.0.1